[Database] implement id generation

+++ backend/ProjectService/Drimstarter.ProjectService/Domain/Category.cs
@@ -5,7 +5,7 @@ public class Category
     public const int NameMinLength = 2;
     public const int NameMaxLength = 50;
 
-    public short Id { get; set; }
+    public long Id { get; set; }
 
     public required string Name { get; init; }
 }
