+++ backend/ProjectService/Drimstarter.ProjectService/Migrations/20240825065613_AddCategory.cs
@@ -15,7 +15,7 @@ namespace Drimstarter.ProjectService.Migrations
                 name: "Categories",
                 columns: table => new
                 {
-                    Id = table.Column<short>(type: "smallint", nullable: false)
+                    Id = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn),
                     Name = table.Column<string>(type: "character varying(50)", maxLength: 50, nullable: false)
                 },
