+++ backend/ProjectService/Drimstarter.ProjectService/Migrations/ProjectDbContextModelSnapshot.cs
@@ -23,11 +23,11 @@ namespace Drimstarter.ProjectService.Migrations
 
             modelBuilder.Entity("Drimstarter.ProjectService.Domain.Category", b =>
                 {
-                    b.Property<short>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("smallint");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<short>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Name")
                         .IsRequired()
